requirements.txt
django
django-htmx
django-extensions
django-filter
django-debug-toolbar
xlwings
pandas
rich